use uintX_2 for jamlib

+++ Envy/jamlib-1.4.7/src/crc32.c
@@ -23,7 +23,7 @@
 #include <jamlib/jam.h>
 
 /* CRC polynomial 0xedb88320 {{{ */
-static ulong crc32tab[] = {
+static uint32_t crc32tab[] = {
     0x00000000, 0x77073096, 0xee0e612c, 0x990951ba, 0x076dc419, 0x706af48f, 0xe963a535, 0x9e6495a3,
     0x0edb8832, 0x79dcb8a4, 0xe0d5e91e, 0x97d2d988, 0x09b64c2b, 0x7eb17cbd, 0xe7b82d07, 0x90bf1d91,
     0x1db71064, 0x6ab020f2, 0xf3b97148, 0x84be41de, 0x1adad47d, 0x6ddde4eb, 0xf4d4b551, 0x83d385c7,
@@ -64,15 +64,15 @@ static ulong crc32tab[] = {
  **
  ** JAM_Crc32 - calcutate CRC32 {{{
  **/
-ulong JAM_Crc32(uchar *Buffer_PC, ulong Length_I)
+uint32_t JAM_Crc32(uint8_t *Buffer_PC, uint32_t Length_I)
 {
-    ulong Crc_I;
+    uint32_t Crc_I;
     int c;
 
     Crc_I=0xffffffff;
 
     for(c=0;c < Length_I; c++)
-	Crc_I=(Crc_I>>8) ^ crc32tab[(uchar)Crc_I ^ tolower(Buffer_PC[c])];
+	Crc_I=(Crc_I>>8) ^ crc32tab[(uint8_t)Crc_I ^ tolower(Buffer_PC[c])];
 
     return(Crc_I);
 }
